# Set proper service name in dedicated client tchannels