--- a/runtime/gateway_test.go
+++ b/runtime/gateway_test.go
@@ -143,6 +143,7 @@ func TestGatewaySetupClientTChannelWhenServiceNameAlreadyExists(t *testing.T) {
 func TestGatewaySetupClientTChannel(t *testing.T) {
 	cfg := NewStaticConfigOrDie(nil, map[string]interface{}{
 		"tchannel.processName": "test-proc",
+		"tchannel.serviceName": "test-gateway",
 	})
 	g := Gateway{
 		TChannelSubLoggerLevel: zapcore.ErrorLevel,