--- a/test/endpoints/baz/baz_metrics_test.go
+++ b/test/endpoints/baz/baz_metrics_test.go
@@ -168,7 +168,7 @@ func TestCallMetrics(t *testing.T) {
 	tchannelTags := map[string]string{
 		"env":             "test",
 		"app":             "test-gateway",
-		"service":         "bazService",
+		"service":         "test-gateway",
 		"target-service":  "bazService",
 		"target-endpoint": "SimpleService__call",
 		"host":            zanzibar.GetHostname(),