## Summary
| PR Status  | Type  | Impact level |
| :---: | :---: | :---: |
| Ready  | Bug | High |



## Description
- Set tchannel service name to be the same as server (gateway) name

## Motivation 
- The newly initiated channels as part of #778 were passed a service name of their client-id while in essence they should use the same service name as the gateway. This is because even though y
- A new boolean flag `dedicated.tchannel.client` is introduced. When set to `true`, each TChannel client creates a dedicated connection instead of registering on the default shared server connection.
Clients need not do anything as the default behaviour is to use a shared connection.([#778](https://github.com/uber/zanzibar/pull/778))


- A new field has been added to the gateway to propagate the `TChannelSubLoggerLevel` for clients with dedicated connections.

---

## Reviewer Notes
- Only 6-7 files need a review. Rest are generated / tests
- Estimated time to review: 20 mins given that you've read the internal ERD :)

## Description
This PR adds / changes the following

- Changes the way Edge `TChannel` Clients connect with the sidecar. Instead of using a common gateway channel and router, we create dedicated channels for each client. This is done based on the flag `dedicated.tchannel.client: true`
![image](https://user-images.githubusercontent.com/12872673/128399975-24509c4d-41fc-4aaa-862b-2d658ec6a71b.png)

- Shutdown hooks now include the shutdown for clients in a separate goroutine.
- Refactor existing gateway channel and router to more readable names
- **Does not** change templates for tchannel endpoints. They still get registered on the common server router. Only clients get new routers.
- Fixes some formatting consistencies and typos.


## Motivation & Context
- Current TChannel clients speak like below
![image](https://user-images.githubusercontent.com/12872673/128399876-330b6703-9b56-44e1-936a-381d587f2298.png)

- Maintaining a dedicated connection ensures that in the event of a misbehaving downstream, requests to other “healthy” tchannel clients do not time out. Only the backlog requests in the queue for the unhealthy service would be affected if the timeout expires. This leads to a more fault tolerant request layer in Edge. As a Tier 0 service, this offers significant platform stability.

- Since each queue would be separate and multiple router-sidecar goroutines would be reading requests in parallel, we expect good improvements in request latency.

- Slight although insignificant improvements may be seen in CPU (more utilization) and memory (reduced pressure of requests in the queue).

## How was this tested?
Tested this version of Zanzibar in Uber's Edge Gateway and verifying that 
- Multiple tchannel clients in Edge get their own channels and are able to make RPCs
- Statsd metrics are emitted
```
tchannel.client.running+client=bazService,dc=unknown,env=test,service=test-gateway
      Value:v2.MetricValue{MetricType:2, Count:0, Gauge:1, Timer:0},..., 
```
- Shutdown hooks are called


Tested that conditional client generation works and when `dedicated.tchannel.client: false | nil`, the existing behaviour remains true




## Benchmarks


**Verdict:** No considerable change in op/sec. Zero change in memory allocs

### Branch
```

========

Branch
time -p sh -c "go test -run _NONE_ -bench . -benchmem -benchtime 7s -cpu 2 ./test/... | grep -v '^ok ' | grep -v '\[no test files\]' | grep -v '^PASS'"
goos: darwin
goarch: amd64
pkg: github.com/uber/zanzibar/test
cpu: Intel(R) Core(TM) i7-9750H CPU @ 2.60GHz
  108061	     67386 ns/op	   21593 B/op	     165 allocs/op
goos: darwin
goarch: amd64
pkg: github.com/uber/zanzibar/test/endpoints/baz
cpu: Intel(R) Core(TM) i7-9750H CPU @ 2.60GHz
   36069	    232823 ns/op	   85245 B/op	     716 allocs/op
   36259	    229871 ns/op	   80946 B/op	     713 allocs/op
   39594	    213469 ns/op	   74620 B/op	     671 allocs/op
   38223	    213605 ns/op	   74171 B/op	     657 allocs/op
goos: darwin
goarch: amd64
pkg: github.com/uber/zanzibar/test/endpoints/contacts
cpu: Intel(R) Core(TM) i7-9750H CPU @ 2.60GHz
   42102	    190982 ns/op	   73774 B/op	     559 allocs/op
goos: darwin
goarch: amd64
pkg: github.com/uber/zanzibar/test/endpoints/googlenow
cpu: Intel(R) Core(TM) i7-9750H CPU @ 2.60GHz
   52650	    155615 ns/op	   63501 B/op	     441 allocs/op
real 96.87
user 121.16
sys 39.36
========
````

### Master

```
========
time -p sh -c "go test -run _NONE_ -bench . -benchmem -benchtime 7s -cpu 2 ./test/... | grep -v '^ok ' | grep -v '\[no test files\]' | grep -v '^PASS'"
goos: darwin
goarch: amd64
pkg: github.com/uber/zanzibar/test
cpu: Intel(R) Core(TM) i7-9750H CPU @ 2.60GHz
  126316	     56844 ns/op	   21978 B/op	     165 allocs/op
goos: darwin
goarch: amd64
pkg: github.com/uber/zanzibar/test/endpoints/baz
cpu: Intel(R) Core(TM) i7-9750H CPU @ 2.60GHz
   35053	    228300 ns/op	   88222 B/op	     716 allocs/op
   36102	    226970 ns/op	   83794 B/op	     713 allocs/op
   39868	    213884 ns/op	   76952 B/op	     671 allocs/op
   39051	    214590 ns/op	   76357 B/op	     657 allocs/op
goos: darwin
goarch: amd64
pkg: github.com/uber/zanzibar/test/endpoints/contacts
cpu: Intel(R) Core(TM) i7-9750H CPU @ 2.60GHz
   42651	    195224 ns/op	   78067 B/op	     559 allocs/op
goos: darwin
goarch: amd64
pkg: github.com/uber/zanzibar/test/endpoints/googlenow
cpu: Intel(R) Core(TM) i7-9750H CPU @ 2.60GHz
   51500	    151183 ns/op	   60568 B/op	     441 allocs/op
real 98.17
user 129.94
sys 42.91
========
```